pyproject.toml
[project]
name = "llmindex"
version = "0.1.0"
description = "LLM-ready index generator for websites — spec, validator, and CLI tools"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "llmindex contributors"},
]
keywords = ["llm", "ai", "search", "seo", "manifest", "schema", "json", "structured-data"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2.6",
    "jsonschema>=4.21",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/mci77777/openllmindex"
Documentation = "https://github.com/mci77777/openllmindex#readme"
Repository = "https://github.com/mci77777/openllmindex"
Issues = "https://github.com/mci77777/openllmindex/issues"
Changelog = "https://github.com/mci77777/openllmindex/releases"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]
templates = [
    "jinja2>=3.1",
]
verify = [
    "dnspython>=2.6",
    "httpx>=0.27",
]

[project.scripts]
llmindex = "llmindex.llmindex_cli.main:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["llmindex*", "spec*"]

[tool.setuptools.package-data]
spec = ["schemas/*.json", "examples/**/*", "test-vectors/**/*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["llmindex/tests"]
addopts = "--cov=llmindex --cov-report=term-missing --cov-fail-under=85"
